requirements.txt
# Core dependencies
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0

# Document processing
PyMuPDF>=1.23.0  # For PDF extraction (fitz)
python-docx>=1.1.0  # For DOCX extraction
chardet>=5.2.0  # For text encoding detection

# Vector stores
faiss-cpu>=1.7.4  # FAISS for vector search
chromadb>=0.4.22  # Alternative vector store

# Utilities
numpy>=1.24.0  # For embeddings manipulation
tiktoken>=0.5.2  # For token counting (optional)
